- moved algorithms test clients
